Add version compatibility ranges for dependencies (#1115)
* Add dependency ranges

Without dependency ranges specified in the pyproject, the version of each
dependency actually installed is hard to predict. This can sometimes lead to
bugs, like the one I just encountered where sessions could not be loaded if
`harp-python` is version 0.2. I have added dependency ranges for each of the
project's dependencies, which should include the current released version and
all subsequent versions that are compatible with the current version, according
to semver. I have left the pinned versions as is, as they are likely not
following semver.

* Remove conda requirements.txt file

There should be a single source of truth, and it seems that this file is no
longer being used. It is redundant with the dependencies of the pyproject.toml.